improvements

+++ src/service/remove_user.py
@@ -1,21 +1,30 @@
+import subprocess
+
 from src.util import connection as con
 from src.service.select_users import SelectUser
+from src.service.sign_in import SignIn
 
 
 class RemoveUser:
 
     def __init__(self):
 
-        SelectUser()
+        if SignIn.get_user_type() == 'admin':
 
-        print('\nDelete user\n')
+            SelectUser()
 
-        print('\n' + '\033[31m' + '0: Cancel' + '\033[0;0m')
-        user_id = int(input('Type the user id: '))
+            print('\nDelete user')
 
-        if user_id == 0:
-            #amaisen.start()
-            None
+            print('\n' + '\033[31m' + '0: Cancel' + '\033[0;0m')
+            user_id = int(input('Type the user id: '))
+
+            if user_id == 0:
+                None
+            else:
+                query = ('DELETE FROM USER WHERE ID = %i' % user_id)
+                con.remove_data(query)
+
         else:
-            query = ('DELETE FROM USER WHERE ID = %i' % user_id)
-            con.remove_data(query)
+            subprocess.run(['clear'])
+            print('\033[31m' + 'You have no permissions to do it!' + '\033[0;0m')
+            print(input('Press any key to continue...'))
